+++ src/types/index.ts
@@ -4,3 +4,4 @@ export * from './widget';
 export * from './insights';
 export * from './algoliasearch';
 export * from './middleware';
+export * from './utils';